requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.11
django-braces==1.13.0
django-crispy-forms==1.7.2
django-mysql==2.3.0
django-nocaptcha-recaptcha==0.0.20
idna==2.6
pytz==2018.3
requests==2.18.4
selenium==3.12.0
urllib3==1.22
